django/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
distlib==0.3.1
dj-database-url==0.4.2
Django==3.1.7
django-cors-headers==3.7.0
django-markdown==0.8.4
django-markdownx==3.0.1
django-widget-tweaks==1.4.1
django-widgets-improved==1.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docker==4.4.4
filelock==3.0.12
hiredis==1.1.0
httpie==2.4.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
Markdown==3.3.4
msgpack==1.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
Pillow==8.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
PySocks==1.7.1
python-decouple==3.1
pytz==2021.1
pywin32==227
requests==2.25.1
requests-toolbelt==0.9.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.3
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
websocket-client==0.58.0
whitenoise==5.2.0
zope.interface==5.2.0
